Move label common back to label_like
